+++ src/adservice/src/main/java/hipstershop/AdService.java
@@ -21,9 +21,9 @@ import com.google.common.collect.Iterables;
 import hipstershop.Demo.Ad;
 import hipstershop.Demo.AdRequest;
 import hipstershop.Demo.AdResponse;
-import io.grpc.Server;
-import io.grpc.ServerBuilder;
-import io.grpc.StatusRuntimeException;
+import hipstershop.Demo.GetFlagResponse;
+import hipstershop.FeatureFlagServiceGrpc.FeatureFlagServiceBlockingStub;
+import io.grpc.*;
 import io.grpc.health.v1.HealthCheckResponse.ServingStatus;
 import io.grpc.protobuf.services.*;
 import io.grpc.stub.StreamObserver;
@@ -62,24 +62,40 @@ public final class AdService {
   private static final Tracer tracer = GlobalOpenTelemetry.getTracer("adservice");
   private static final Meter meter = GlobalOpenTelemetry.getMeter("adservice");
 
-  private static final LongCounter adRequestsCounter = meter
+  private static final LongCounter adRequestsCounter =
+      meter
           .counterBuilder("app.ads.ad_requests")
           .setDescription("Counts ad requests by request and response type")
           .build();
 
-  private static final AttributeKey<String> adRequestTypeKey = AttributeKey.stringKey("app.ads.ad_request_type");
-  private static final AttributeKey<String> adResponseTypeKey = AttributeKey.stringKey("app.ads.ad_response_type");
+  private static final AttributeKey<String> adRequestTypeKey =
+      AttributeKey.stringKey("app.ads.ad_request_type");
+  private static final AttributeKey<String> adResponseTypeKey =
+      AttributeKey.stringKey("app.ads.ad_response_type");
 
   private void start() throws IOException {
-    int port = Integer.parseInt(Optional.ofNullable(System.getenv("AD_SERVICE_PORT")).orElseThrow(
-        () -> new IOException(
-            "environment vars: AD_SERVICE_PORT must not be null")
-    ));
+    int port =
+        Integer.parseInt(
+            Optional.ofNullable(System.getenv("AD_SERVICE_PORT"))
+                .orElseThrow(
+                    () ->
+                        new IllegalStateException(
+                            "environment vars: AD_SERVICE_PORT must not be null")));
     healthMgr = new HealthStatusManager();
+
+    String featureFlagServiceAddr =
+        Optional.ofNullable(System.getenv("FEATURE_FLAG_GRPC_SERVICE_ADDR"))
+            .orElseThrow(
+                () ->
+                    new IllegalStateException(
+                        "environment vars: FEATURE_FLAG_GRPC_SERVICE_ADDR must not be null"));
+    FeatureFlagServiceBlockingStub featureFlagServiceStub =
+        FeatureFlagServiceGrpc.newBlockingStub(
+            ManagedChannelBuilder.forTarget(featureFlagServiceAddr).usePlaintext().build());
 
     server =
         ServerBuilder.forPort(port)
-            .addService(new AdServiceImpl())
+            .addService(new AdServiceImpl(featureFlagServiceStub))
             .addService(healthMgr.getHealthService())
             .build()
             .start();
@@ -105,14 +121,24 @@ public final class AdService {
   }
 
   private enum AdRequestType {
-    TARGETED, NOT_TARGETED
+    TARGETED,
+    NOT_TARGETED
   }
 
   private enum AdResponseType {
-    TARGETED, RANDOM
+    TARGETED,
+    RANDOM
   }
 
   private static class AdServiceImpl extends hipstershop.AdServiceGrpc.AdServiceImplBase {
+
+    private static final String ADSERVICE_FAIL_FEATURE_FLAG = "adServiceFailure";
+
+    private final FeatureFlagServiceBlockingStub featureFlagServiceStub;
+
+    private AdServiceImpl(FeatureFlagServiceBlockingStub featureFlagServiceStub) {
+      this.featureFlagServiceStub = featureFlagServiceStub;
+    }
 
     /**
      * Retrieves ads based on context provided in the request {@code AdRequest}.
@@ -156,7 +182,15 @@ public final class AdService {
         span.setAttribute("app.ads.ad_request_type", adRequestType.name());
         span.setAttribute("app.ads.ad_response_type", adResponseType.name());
 
-        adRequestsCounter.add(1, Attributes.of(adRequestTypeKey, adRequestType.name(), adResponseTypeKey, adResponseType.name()));
+        adRequestsCounter.add(
+            1,
+            Attributes.of(
+                adRequestTypeKey, adRequestType.name(), adResponseTypeKey, adResponseType.name()));
+
+        if (checkAdFailure()) {
+          logger.warn(ADSERVICE_FAIL_FEATURE_FLAG + " fail feature flag enabled");
+          throw new StatusRuntimeException(Status.RESOURCE_EXHAUSTED);
+        }
 
         AdResponse reply = AdResponse.newBuilder().addAllAds(allAds).build();
         responseObserver.onNext(reply);
@@ -168,6 +202,20 @@ public final class AdService {
         logger.log(Level.WARN, "GetAds Failed with status {}", e.getStatus());
         responseObserver.onError(e);
       }
+    }
+
+    boolean checkAdFailure() {
+      // Flip a coin and fail 1/10th of the time if feature flag is enabled
+      if (random.nextInt(10) != 1) {
+        return false;
+      }
+
+      GetFlagResponse response =
+          featureFlagServiceStub.getFlag(
+              hipstershop.Demo.GetFlagRequest.newBuilder()
+                  .setName(ADSERVICE_FAIL_FEATURE_FLAG)
+                  .build());
+      return response.getFlag().getEnabled();
     }
   }
 
@@ -259,7 +307,7 @@ public final class AdService {
         .putAll("accessories", colorImager, solarFilter, cleaningKit)
         .putAll("assembly", opticalTube)
         .putAll("travel", travelTelescope)
-            // Keep the books category free of ads to ensure the random code branch is tested
+        // Keep the books category free of ads to ensure the random code branch is tested
         .build();
   }