+++ CHANGELOG.md
@@ -200,6 +200,8 @@ significant modifications will be credited to OpenTelemetry Authors.
 ([#693](https://github.com/open-telemetry/opentelemetry-demo/pull/693))
 * Update recommendationservice python base image and dependencies
 ([#700](https://github.com/open-telemetry/opentelemetry-demo/pull/700))
+* Add adServiceFailure feature flag triggering Ad Service errors
+([#694](https://github.com/open-telemetry/opentelemetry-demo/pull/694))
 * Reduce spans generated from quote service
 ([#702](https://github.com/open-telemetry/opentelemetry-demo/pull/702))
 * Update emailservice Dockerfile to use alpine and multistage build